Render a sphere with normals
I think it's more fun for me if I don't port the original code directly, but instead build it up.

One that that happened was that he added, unusually, a way to get the nearest point in the scene. However, he doesn't use that info elsewhere, so I wanted to experiment with using it (for something like getting the normal calculation)